# contracts/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: API SysAP
  version: 0.2.0
  description: >-
    Contrato HTTP do SysAP. Saude e prontidao estao implementadas; a superficie
    versionada de identidade descreve o contrato planejado para a Fase 2 e nao
    representa funcionalidade disponivel nesta subfase documental.
servers:
  - url: http://127.0.0.1:8080
    description: API local em loopback
security:
  - BearerAccessToken: []
...
paths:
...
  /v1/organizations/{organization_id}/athlete-invitations:
    post:
      operationId: createAthleteInvitation
      tags:
        - Convites de atleta
      summary: Pre-cadastrar atleta e iniciar ativacao
      description: >-
        Cria um convite na organizacao selecionada por organization_id.
        Requer token de Owner.
      x-sysap-implementation-status: planned
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OrganizationContext"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAthleteInvitationRequest"
            examples:
              fictionalAthlete:
                summary: Atleta ficticio
                value:
                  display_name: "Atleta Exemplo"
                  phone_e164: "+12025550123"
                  email: "atleta@example.com"
      responses:
        "201":
          description: Convite criado
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteInvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "409":
          $ref: "#/components/responses/ResourceConflict"
        "500":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/activation/start:
    post:
      operationId: startActivation
      tags:
        - Convites de atleta
      summary: Iniciar ativação de atleta
      description: >-
        Inicia o fluxo de ativação, gerando um OTP para a matrícula informada.
        Sempre retorna 202 Accepted para não enumerar matrículas válidas.
        Em desenvolvimento, o provedor é local; em produção, o envio é
        configurado no Supabase Auth. A API não integra provedores SMS diretamente.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationStartRequest"
      responses:
        "202":
          description: Solicitação aceita
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: accepted
        "503":
          $ref: "#/components/responses/IdentityVerificationUnavailable"
  /v1/activation/verify:
    post:
      operationId: verifyActivation
      tags:
        - Convites de atleta
      summary: Verificar OTP
      description: >-
        Verifica se o OTP fornecido corresponde à matrícula e não expirou.
        Retorna uma prova de ativação (activation_proof) que pode ser usada depois.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationVerifyRequest"
      responses:
        "200":
          description: OTP validado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationVerifyResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "503":
          $ref: "#/components/responses/IdentityVerificationUnavailable"
  /v1/activation/complete:
    post:
      operationId: completeActivation
      tags:
        - Convites de atleta
      summary: Concluir ativação
      description: >-
        Conclui o fluxo de ativação, validando a activation_proof e definindo a senha do atleta.
        Provisiona a identidade pelo adaptador server-side do Supabase Auth.
      x-sysap-implementation-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationCompleteRequest"
      responses:
        "200":
          description: Ativação concluída com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationCompleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "409":
          $ref: "#/components/responses/ResourceConflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/auth/login:
    post:
      operationId: login
      tags:
        - Autenticacao
      summary: Autenticar por matrícula e senha
      description: >-
        Autentica athlete, trainer ou owner com matrícula e senha. A API é a
        única intermediária com o Supabase Auth; antes de emitir tokens ela
        confirma perfil, organização e membership no estado autoritativo local.
        Matrícula inexistente, senha inválida, suspensão, ativação pendente e
        bloqueio temporário retornam exatamente a mesma resposta 401.
      x-sysap-implementation-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Credenciais aceitas e sessao emitida.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
...
  /v1/me:
    get:
      operationId: getCurrentIdentity
      tags:
        - Identidade
      summary: Consultar identidade atual
      description: >-
        Retorna perfil e memberships confirmadas no estado autoritativo do
        SysAP. O resultado nao expoe identificador tecnico de autenticacao,
        telefone, tokens nem claims usados como unica fonte de autorizacao.
      x-sysap-implementation-status: planned
      security:
        - BearerAccessToken: []
      responses:
        "200":
          description: Identidade e vinculos atuais.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentIdentityResponse"
              examples:
                fictionalAthlete:
                  summary: Perfil ficticio de atleta
                  value:
                    profile:
                      id: 40000000-0000-4000-8000-000000000001
                      display_name: Atleta Exemplo
                    memberships:
                      - organization_id: 30000000-0000-4000-8000-000000000001
                        role: athlete
                        status: active
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/IdentityVerificationUnavailable"
...
components:
  securitySchemes:
    BearerAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Access token de curta duracao validado pela API. O token e tratado pelo
        cliente como opaco; sub/session_id nao sao IDs de negocio. Sessao deve
        ter sido registrada pela API; papel, organizacao e acesso sao
        confirmados no estado do SysAP.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        UUID gerado pelo cliente para tornar o comando repetivel. Comandos sem
        segredo comparam somente a projecao canonica nao secreta. Em ativacao e
        recovery, desafio/estado terminal controlam repeticao; senha, OTP/TOTP,
        token, ticket MFA e qualquer derivado nunca entram no fingerprint.
      schema:
        type: string
        format: uuid
      example: 00000000-0000-4000-8000-000000000001
    OrganizationContext:
      name: X-Organization-ID
      in: header
      required: true
      description: >-
        Seletor sujeito a autorizacao, obrigatorio somente para endpoints
        tenant-bound futuros. Nao e exigido em /healthz, /readyz ou futuro
        /v1/me. Um formato invalido ou malformado produz invalid_request.
      schema:
        type: string
        format: uuid
      example: 30000000-0000-4000-8000-000000000001
...
  responses:
...
    BadRequest:
      description: Requisicao malformada ou header invalido.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            malformedRequest:
              summary: Requisicao malformada
              value:
                error:
                  code: invalid_request
                  message: request is invalid
                  request_id: 0123456789abcdef0123456789abcdef
    InvalidCredentials:
      description: Falha generica de autenticacao sem indicar qual dado falhou.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            authenticationFailed:
              summary: Autenticacao nao concluida
              value:
                error:
                  code: invalid_credentials
                  message: invalid credentials
                  request_id: 0123456789abcdef0123456789abcdef
    AuthenticationRequired:
      description: >-
        Access token ausente, invalido ou expirado, ou sessao local inexistente
        ou revogada.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        WWW-Authenticate:
          $ref: "#/components/headers/BearerChallenge"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            authenticationRequired:
              summary: Autenticacao necessaria
              value:
                error:
                  code: authentication_required
                  message: authentication is required
                  request_id: 0123456789abcdef0123456789abcdef
    AccessDenied:
      description: Sujeito autenticado sem acesso ativo ou permissao suficiente.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            accessDenied:
              summary: Acesso negado
              value:
                error:
                  code: access_denied
                  message: access is denied
                  request_id: 0123456789abcdef0123456789abcdef
    IdentityVerificationUnavailable:
      description: >-
        Nenhuma chave JWKS confiavel esta disponivel para verificacao de
        identidade. Esta resposta e planejada para GET /v1/me e nao revela
        detalhes do token, claim, emissor ou infraestrutura.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            identityVerificationUnavailable:
              summary: Verificacao de identidade indisponivel
              value:
                error:
                  code: identity_verification_unavailable
                  message: identity verification is temporarily unavailable
                  request_id: 0123456789abcdef0123456789abcdef
...
    ResourceConflict:
      description: Estado incompativel ou chave de idempotencia reutilizada.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            resourceConflict:
              summary: Conflito de estado
              value:
                error:
                  code: resource_conflict
                  message: request conflicts with current state
                  request_id: 0123456789abcdef0123456789abcdef
    ValidationFailed:
      description: Campos invalidos sem ecoar valores sensiveis recebidos.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          examples:
            validationFailed:
              summary: Falha de validacao estatica
              value:
                error:
                  code: validation_failed
                  message: request validation failed
                  request_id: 0123456789abcdef0123456789abcdef
                  details:
                    - field: enrollment_number
                      reason: invalid_format
    RateLimited:
      description: Limite temporario excedido sem revelar a dimensao aplicada.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            rateLimited:
              summary: Limite temporario excedido
              value:
                error:
                  code: rate_limited
                  message: too many requests
                  request_id: 0123456789abcdef0123456789abcdef
    ServiceUnavailable:
      description: >-
        Operacao temporariamente indisponivel sem identificar dependencia ou
        fornecedor. Em fluxos antienumeracao, somente falha global usa 503.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            serviceUnavailable:
              summary: Servico temporariamente indisponivel
              value:
                error:
                  code: service_unavailable
                  message: service is temporarily unavailable
                  request_id: 0123456789abcdef0123456789abcdef
  headers:
...
    RequestID:
      description: >-
        Identificador hexadecimal gerado pelo servidor para correlacao da
        requisicao. O caminho normal possui 32 caracteres; a contingencia pode
        produzir um valor hexadecimal maior.
      required: true
      schema:
        $ref: "#/components/schemas/RequestID"
      example: 0123456789abcdef0123456789abcdef
    NoStore:
      description: Impede armazenamento de respostas de identidade e autenticacao.
      required: true
      schema:
        type: string
        const: no-store
      example: no-store
    RetryAfter:
      description: Segundos inteiros antes de uma nova tentativa permitida.
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 86400
      example: 60
    BearerChallenge:
      description: Desafio padrao para access token bearer.
      required: true
      schema:
        type: string
        const: Bearer
      example: Bearer
...
  schemas:
    RequestID:
      type: string
      pattern: "^[0-9a-f]+$"
      minLength: 32
      description: Identificador hexadecimal opaco gerado pelo servidor.
...
    UUID:
      type: string
      format: uuid
      description: Identificador opaco no dominio do SysAP.
    EnrollmentNumber:
      type: string
      minLength: 10
      maxLength: 10
      pattern: "^[0-9]{10}$"
      description: Matricula publica ficticia de exatamente dez digitos.
      example: "2026000001"
...
    PasswordInput:
      type: string
      minLength: 15
      maxLength: 128
      writeOnly: true
      description: >-
        Senha escolhida pelo usuario. Espacos e Unicode sao permitidos, sem
        truncamento silencioso nem regra previsivel de composicao.
    OTPInput:
      type: string
      minLength: 6
      maxLength: 6
      pattern: "^[0-9]{6}$"
      writeOnly: true
      description: OTP numerico de seis digitos, tratado como segredo de uso unico.
...
    AccessTokenOutput:
      type: string
      minLength: 1
      maxLength: 16384
      readOnly: true
      description: Access token de curta duracao retornado somente apos autenticacao.
    RefreshTokenOutput:
      type: string
      minLength: 1
      maxLength: 16384
      readOnly: true
      description: Refresh token rotacionavel retornado somente apos autenticacao.
...
    MembershipRole:
      type: string
      enum:
        - owner
        - trainer
        - athlete
      description: Papel confirmado no estado autoritativo do SysAP.
    MembershipAccessStatus:
      type: string
      enum:
        - active
        - suspended
      description: >-
        Projecao mutavel de acesso. Membership pending_activation ou closed nao
        aparece em /me e nao pode ser alvo deste PATCH.
    CreateAthleteInvitationRequest:
      type: object
      additionalProperties: false
      required:
        - display_name
        - phone_e164
      properties:
        display_name:
          type: string
          minLength: 1
          description: Nome de exibicao do atleta
        phone_e164:
          type: string
          pattern: '^\+[1-9]\d{1,14}$'
          description: Telefone normalizado em formato E.164
        email:
          type: string
          format: email
          description: Email opcional do atleta
    AthleteInvitationResponse:
      type: object
      additionalProperties: false
      required:
        - invitation_id
        - profile_id
        - enrollment_number
        - status
        - expires_at
      properties:
        invitation_id:
          $ref: "#/components/schemas/UUID"
        profile_id:
          $ref: "#/components/schemas/UUID"
        enrollment_number:
          $ref: "#/components/schemas/EnrollmentNumber"
        status:
          type: string
          enum: [pending_activation]
          description: O atleta deve ativar a conta com senha e OTP.
        expires_at:
          type: string
          format: date-time
          description: Data de expiração do convite.
    ActivationStartRequest:
      type: object
      additionalProperties: false
      required:
        - enrollment_number
      properties:
        enrollment_number:
          $ref: "#/components/schemas/EnrollmentNumber"
    ActivationVerifyRequest:
      type: object
      additionalProperties: false
      required:
        - enrollment_number
        - code
      properties:
        enrollment_number:
          $ref: "#/components/schemas/EnrollmentNumber"
        code:
          $ref: "#/components/schemas/OTPInput"
    ActivationVerifyResponse:
      type: object
      additionalProperties: false
      required:
        - activation_proof
      properties:
        activation_proof:
          type: string
          description: Prova de validação do OTP (hexadecimal, única, opaca)
          example: 1a2b3c4d5e6f7g8h
    ActivationCompleteRequest:
      type: object
      required:
        - activation_proof
        - password
      properties:
        activation_proof:
          type: string
          description: Prova obtida no step de verify
          example: 1a2b3c4d5e6f7g8h
        password:
          type: string
          format: password
          description: Senha do atleta
    ActivationCompleteResponse:
      type: object
      properties:
        profile_id:
          type: string
          format: uuid
        enrollment_number:
          type: string
        role:
          type: string
          example: athlete
        organization_id:
          type: string
          format: uuid
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - enrollment_number
        - password
      properties:
        enrollment_number:
          $ref: "#/components/schemas/EnrollmentNumber"
        password:
          $ref: "#/components/schemas/PasswordInput"
    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - session_id
        - profile_id
        - organization_id
        - role
        - aal
        - access_token
        - refresh_token
        - expires_in
      properties:
        session_id:
          $ref: "#/components/schemas/UUID"
        profile_id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        role:
          $ref: "#/components/schemas/MembershipRole"
        aal:
          type: string
          const: aal1
        access_token:
          $ref: "#/components/schemas/AccessTokenOutput"
        refresh_token:
          $ref: "#/components/schemas/RefreshTokenOutput"
        expires_in:
          type: integer
          minimum: 1
...
    CurrentIdentityResponse:
      type: object
      additionalProperties: false
      required:
        - profile
        - memberships
      properties:
        profile:
          $ref: "#/components/schemas/CurrentProfile"
        memberships:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/CurrentMembership"
    CurrentProfile:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        display_name:
          type: string
          minLength: 1
          maxLength: 120
    CurrentMembership:
      type: object
      additionalProperties: false
      required:
        - organization_id
        - role
        - status
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        role:
          $ref: "#/components/schemas/MembershipRole"
        status:
          $ref: "#/components/schemas/MembershipAccessStatus"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/PublicError"
    PublicError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 64
          pattern: "^[a-z][a-z0-9_]*$"
        message:
          type: string
          minLength: 1
          maxLength: 200
        request_id:
          $ref: "#/components/schemas/RequestID"
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - request_id
        - details
      properties:
        code:
          type: string
          const: validation_failed
        message:
          type: string
          const: request validation failed
        request_id:
          $ref: "#/components/schemas/RequestID"
        details:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/ValidationDetail"
    ValidationDetail:
      type: object
      additionalProperties: false
      required:
        - field
        - reason
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[a-z][a-z0-9_.]*$"
        reason:
          type: string
          enum:
            - required
            - invalid_format
            - too_short
            - too_long
            - not_allowed
            - password_rejected